rename file for migration tests to pass
same thing

added to dto class
