+++ db/migration/V0057__Create_notify_trigger_on_lobby_table.SQL
