+++ src/main/java/com/patina/codebloom/common/components/DuelManager.java
@@ -2,6 +2,7 @@ package com.patina.codebloom.common.components;
 
 import org.springframework.stereotype.Component;
 
+import com.patina.codebloom.common.dto.lobby.DuelData;
 import com.patina.codebloom.common.dto.lobby.LobbyDto;
 import com.patina.codebloom.common.db.models.lobby.Lobby;
 import com.patina.codebloom.common.db.repos.lobby.LobbyRepository;
